[refactor] store.js에서 reducer.js를 분리
