+++ src/reducer.js
@@ -1,0 +1,42 @@
+const initialState = {
+  name: '',
+  category: '',
+  address: '',
+  data: [],
+};
+
+export default function reducer(state = initialState, action) {
+  if (action.type === 'handleChangeNameValue') {
+    return {
+      ...state,
+      name: action.payload.name,
+    };
+  }
+  if (action.type === 'handleChangeCategroryValue') {
+    return {
+      ...state,
+      category: action.payload.category,
+    };
+  }
+  if (action.type === 'handleChangeAddressValue') {
+    return {
+      ...state,
+      address: action.payload.address,
+    };
+  }
+  if (action.type === 'saveData') {
+    console.log(action.payload.data);
+    return {
+      ...state,
+      name: '',
+      category: '',
+      address: '',
+      data: [
+        ...state.data,
+        action.payload.data,
+      ],
+    };
+  }
+
+  return state;
+}
